requirements.txt
Django>=5
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ipdb==0.13.13
Pillow==10.1.0
dj-rest-auth[with-social]
